requirements.txt
asgiref==3.8.1
boto3==1.38.34
botocore==1.38.34
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.1
django-storages==1.14.6
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.3
s3transfer==0.13.0
scipy==1.15.3
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.146
ultralytics-thop==2.0.14
urllib3==2.4.0
